+++ khoros/studio/base.py
@@ -14,13 +14,13 @@ from ..utils import core_utils
 def sdk_installed():
     """This function checks to see if the Lithium SDK is installed.
 
-    .. versionadded:: 2.6.0
+    .. versionadded:: 2.5.1
 
     :returns: Boolean value indicating whether or not the Lithium SDK is installed
     """
     try:
         output = core_utils.run_cmd('li')
-        is_installed = True if output.returncode == 0 else False
+        is_installed = True if output['return_code'] == 0 else False
     except FileNotFoundError:
         is_installed = False
     return is_installed
@@ -29,13 +29,13 @@ def sdk_installed():
 def get_sdk_version():
     """This function identifies the currently installed version of the Lithium SDK.
 
-    .. versionadded:: 2.6.0
+    .. versionadded:: 2.5.1
 
     :returns: The SDK version in string format or ``None`` if not installed
     """
     version = None
-    if npm_installed():
-        output = core_utils.run_cmd('npm list -g lithium-sdk').stdout
+    if npm_installed() and sdk_installed():
+        output = core_utils.run_cmd('npm list -g lithium-sdk', decode_output=True, strip_output=False)['stdout']
         output = core_utils.decode_binary(output)
         if '(empty)' not in output:
             version = output.split('lithium-sdk@')[1].split(' ')[0]
@@ -45,7 +45,7 @@ def get_sdk_version():
 def node_installed():
     """This function checks whether or not Node.js is installed.
 
-    .. versionadded:: 2.6.0
+    .. versionadded:: 2.5.1
 
     :returns: Boolean value indicating whether or not Node.js is installed
     """
@@ -56,12 +56,14 @@ def node_installed():
 def get_node_version():
     """This function identifies and returns the installed Node.js version.
 
-    .. versionadded:: 2.6.0
+    .. versionadded:: 2.5.1
 
     :returns: The version as a string or ``None`` if not installed
     """
     try:
-        version = core_utils.run_cmd('node -v').stdout.decode('utf-8').strip()[1:]
+        version = core_utils.run_cmd('node -v', decode_output=True, strip_output=True)['stdout']
+        if 'v' in version:
+            version = version[1:]
     except FileNotFoundError:
         version = None
     return version
@@ -70,7 +72,7 @@ def get_node_version():
 def npm_installed():
     """This function checks whether or not npm is installed.
 
-    .. versionadded:: 2.6.0
+    .. versionadded:: 2.5.1
 
     :returns: Boolean value indicating whether or not npm is installed
     """
@@ -81,12 +83,14 @@ def npm_installed():
 def get_npm_version():
     """This function identifies and returns the installed npm version.
 
-    .. versionadded:: 2.6.0
+    .. versionadded:: 2.5.1
 
     :returns: The version as a string or ``None`` if not installed
     """
     try:
-        version = core_utils.run_cmd('npm -v').stdout.decode('utf-8').strip()[1:]
+        version = core_utils.run_cmd('npm -v', decode_output=True, strip_output=True)['stdout']
+        if 'v' in version:
+            version = version[1:]
     except FileNotFoundError:
         version = None
     return version